chore(tanstack-start): remove unreachable RootLayout export
RootLayout was only referenced by its own barrel; the admin html shell
comes from PayloadAdminShell and the provider tree from payloadLayoutRoute.
Moves RootLayoutData to getLayoutData.ts, its only consumer.
