+++ packages/tanstack-start/src/exports/layouts.ts
@@ -1,5 +1,3 @@
 export { getLayoutData } from '../layouts/Root/getLayoutData.js'
-export type { GetLayoutDataArgs } from '../layouts/Root/getLayoutData.js'
-export { RootLayout } from '../layouts/Root/index.js'
-export type { RootLayoutData, RootLayoutProps } from '../layouts/Root/index.js'
+export type { GetLayoutDataArgs, RootLayoutData } from '../layouts/Root/getLayoutData.js'
 export { loadLayoutData } from '../utilities/loadLayoutData.js'
