+++ packages/tanstack-start/src/layouts/Root/getLayoutData.ts
@@ -1,5 +1,14 @@
-import type { AcceptedLanguages } from '@payloadcms/translations'
-import type { ImportMap, LanguageOptions, SanitizedConfig, ServerProps } from 'payload'
+import type { AcceptedLanguages, I18nClient } from '@payloadcms/translations'
+import type { Theme } from '@payloadcms/ui'
+import type {
+  ClientConfig,
+  ImportMap,
+  LanguageOptions,
+  SanitizedConfig,
+  SanitizedPermissions,
+  ServerProps,
+  User,
+} from 'payload'
 
 import { getNavPrefs } from '@payloadcms/ui/elements/Nav/getNavPrefs'
 import { RenderServerComponent } from '@payloadcms/ui/elements/RenderServerComponent'
@@ -9,9 +18,28 @@ import { Outlet } from '@tanstack/react-router'
 import { applyLocaleFiltering } from 'payload/shared'
 import { createElement } from 'react'
 
-import type { RootLayoutData } from './index.js'
-
 import { initReq } from '../../utilities/initReq.server.js'
+
+export type RootLayoutData = {
+  clientConfig: ClientConfig
+  dateFNSKey: I18nClient['dateFNSKey']
+  fallbackLang: string
+  isNavOpen: boolean
+  languageCode: string
+  languageOptions: LanguageOptions
+  locale?: string
+  permissions: SanitizedPermissions
+  /**
+   * Custom admin provider tree (`config.admin.components.providers`) nested
+   * around the router `<Outlet />`. Built unrendered by `getLayoutData`; the
+   * layout server function renders it to an RSC payload before it reaches the
+   * client. `undefined` when no custom providers are configured.
+   */
+  providers?: React.ReactNode
+  theme: Theme
+  translations: I18nClient['translations']
+  user: null | User
+}
 
 export type GetLayoutDataArgs = {
   configPromise: Promise<SanitizedConfig> | SanitizedConfig
